characterizations of (quasi-)separatedness
The aim of this commit is to incorporate the result "(quasi-)separatedness is local on the target" and propositions (1.2.6) and (1.2.7) of EGA IV, première partie (they are in pp. 227-228). (All of them into the section Separation axioms on the Schemes chapter.)

The first one is quasi-separated-local-target. The proof introduces a lot of single-use definitions regarding properties of morphisms of schemes, and actually shows a general result for these definitions (P is LOCT implies Δ_P is LOCT). Such a phrasing for a proof might look weird. I would have written a new section instead on the Schemes chapter with the new definitions plus the mentioned result, but there's been objections to this: https://stacks.math.columbia.edu/tag/01KH#comment-7560

I also changed slightly the statement of lemma-closed-local-target and the statement and proof of lemma-quasi-compact-affine to have the results "«being quasi-compact» is LOCT" and "«being a closed immersion» is LOCT" fully spelled.

Propositions (1.2.6) and (1.2.7) of EGA IV correspond to lemma-characterize-quasi-separated-scheme and lemma-quasi-separatedness-is-topological. In their statements, I've introduced the modern terminology of "quasi-separated topological space" and "quasi-separated map of topological spaces" (EGA doesn't use them), defined in a newly-added section in Topology. There's been hesitations to the introduction of such a section before (see comments in tag 0067). Nonetheless, I thought that giving the concept a name and using it makes more manifest the topological nature of quasi-separatedness. If you think this is not a sufficient reason to include the new section in Topology (I understand the reasons in your comment in tag 0067), one could instead fully spell out the concept each time it appears in lemma-characterize-quasi-separated-scheme and lemma-quasi-separatedness-is-topological.

(Note: the statement of Proposition (1.2.7) of EGA IV contains an erratum, see https://math.stackexchange.com/q/4716228/394668 )