Added definition of a previously undefined notion
The purpose of this commit is to add the previously omitted argument in the last sentence of the proof of lemma-normal-normalization (Chapter on morphisms). For this, I defined in the topology chapter what does it mean for a continuous map of spaces to "induce a bijection on irreducible components." This notion is used undefined for instance in 037A, and in lemma-normal-normalization itself (although its definition is somehow acknowledged in https://stacks.math.columbia.edu/tag/04KX#comment-7268). I also added the obligatory characterization of the notion for sober spaces, lemma-characterize-bijection-irreducible-components-sober.

(We also add the reference to "by a topological argument which we omit" in the proof of lemma-normal-normalization.)