Glueing infinite rank vector bundles
The lemma in this commit tells us that the property of being "locally
projective" for a module on a Noetherian scheme, descends through
finite pushouts... This is something used without mention in the paper
"Locally free twisted sheaves of infinite rank". It's also easy to
prove as this commit shows.