Effective descent
Thanks to Laurent Moret-Bailly who writes

I am enclosing a proposed modified version of Section 39.25 (0CCH) on
effective descent of quasi-projective schemes.

The new Lemma 25.1 (aka {lemma-descend-along-finite-with-bound}) is a
modification of Lemma 0CCI ( with little change in the proof) where the
number of points required to have an affine neighborhood is specified.

This immediately implies the original 0CCI (now Lemma 25.2). There are
no changes to Lemma 25.3 (the usual quasiprojective-like cases).

The reason I am interested is the following Lemma 25.4, stating that
effectiveness always holds for a morphism X-->Y which is finite, locally
free, surjective and radicial.